Create The Miami Marina Murder
Case #003: The Miami Marina Murder
A body was found floating near the docks of Coral Bay Marina in the early hours of August 14, 1986. Your job detective is to find the murderer and bring them to justice. This case might require the use of JOINs, wildcard searches, and logical deduction. Get to work, detective.

Objectives
1.
Find the murderer. ( Start by finding the crime scene and go from there )